+++ src/WebApiBolierplate/API/Operations/Interfaces/IUserLib.cs
@@ -4,7 +4,5 @@ namespace API.Operations.Interfaces;
 
 public interface IUserLib
 {
-    public int ValidateLogin(LoginDTO login);
-
     public User GetUser(string userName);
 }