+++ src/WebApiBolierplate/Core.Lib/Utilities/Interfaces/IJwtUtils.cs
@@ -1,9 +1,8 @@
-﻿using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
+﻿using System.Collections.Generic;
 
-namespace API.Helpers.Interfaces;
+namespace Core.Lib.Utilities.Interfaces;
 
-public interface IJwtTokenBuilder
+public interface IJwtUtils
 {
     public string GenerateToken();
 
@@ -12,12 +11,12 @@ public interface IJwtTokenBuilder
     /// <summary>
     /// Adds the security key used for signing JWT token
     /// </summary>
-    public IJwtTokenBuilder AddSecurityKey(string securityKey);
+    public IJwtUtils AddSecurityKey(string securityKey);
 
     /// <summary>
     /// Adds expiry time in days to JWT token
     /// </summary>
-    public IJwtTokenBuilder AddExpiry(int expiryInDays);
+    public IJwtUtils AddExpiry(int expiryInDays);
 
     #endregion [Adding values]
 
@@ -26,38 +25,38 @@ public interface IJwtTokenBuilder
     /// <summary>
     /// Adds issuer to JWT token
     /// </summary>
-    public IJwtTokenBuilder AddIssuer(string issuer);
+    public IJwtUtils AddIssuer(string issuer);
 
     /// <summary>
     /// Adds audience value to JWT token
     /// </summary>
-    public IJwtTokenBuilder AddAudience(string audience);
+    public IJwtUtils AddAudience(string audience);
 
     /// <summary>
     /// Adds subject as a claim to the token
     /// </summary>
-    public IJwtTokenBuilder AddSubject(string subject);
+    public IJwtUtils AddSubject(string subject);
 
     /// <summary>
     /// Adds user name as claim to the token
     /// </summary>
-    public IJwtTokenBuilder AddName(string username);
+    public IJwtUtils AddName(string username);
 
     /// <summary>
     /// Adds email as claim to JWT token
     /// </summary>
-    public IJwtTokenBuilder AddEmail(string email);
+    public IJwtUtils AddEmail(string email);
 
     /// <summary>
     /// Adds user role as claim to the token
     /// </summary>
-    public IJwtTokenBuilder AddRole(string value);
+    public IJwtUtils AddRole(string value);
 
     #region [Custom Claims]
 
-    public IJwtTokenBuilder AddUserId(string value);
+    public IJwtUtils AddUserId(string value);
 
-    public IJwtTokenBuilder AddCompanyId(string value);
+    public IJwtUtils AddCompanyId(string value);
 
     #endregion [Custom Claims]
 
@@ -69,14 +68,14 @@ public interface IJwtTokenBuilder
     /// <param name="type">claim type</param>
     /// <param name="value">value for the claim</param>
     /// <returns></returns>
-    public IJwtTokenBuilder AddClaim(string type, string value);
+    public IJwtUtils AddClaim(string type, string value);
 
     /// <summary>
     /// Bulk add claims to the JWT tokens
     /// </summary>
     /// <param name="claimList">list of claim types and its corresponding values</param>
     /// <returns></returns>
-    public IJwtTokenBuilder AddClaims(Dictionary<string, string> claimList);
+    public IJwtUtils AddClaims(Dictionary<string, string> claimList);
 
     #endregion [Generic Claims]